split command and change approach
